Fixed the freeze admin in freeze.ts I need to add it to filter.ts

+++ commands/Cooldown/filter.ts
@@ -2,29 +2,30 @@ import discord from "discord.js";
 import { db } from "../../db/db";
 
 export class CooldownFilter {
-  private frozenUsers: IFrozenUser[] = [];
+	private frozenUsers: IFrozenUser[] = [];
 
-  constructor(cli: discord.Client) {
-    cli.on("message", async (message) => {
-      const author = message.author.username.toLowerCase();
-      const frozenUsers = await db.freeze.Get();
+	constructor(cli: discord.Client) {
+		cli.on("message", async (message) => {
+			const author = message.author.username.toLowerCase();
+			const frozenUsers = await db.freeze.Get();
 
-      this.frozenUsers = frozenUsers;
+			this.frozenUsers = frozenUsers;
 
-      for (var i = 0; i != this.frozenUsers.length; i++) {
-        const user = this.frozenUsers[i].username.toLowerCase();
+			for (var i = 0; i != this.frozenUsers.length; i++) {
+				const user = this.frozenUsers[i].username.toLowerCase();
 
-        if (author == user) {
-          if (!message.deletable) return;
+				if (author == user) {
+					if (user == "pinky" || user == "developer" || user == "crew") return;
+					if (!message.deletable) return;
 
-          message.author
-            .createDM()
-            .then((channel) => channel.send("Your Frozen 🥶"))
-            .catch((err) => console.log(`${err.message} ${author}`));
+					message.author
+						.createDM()
+						.then((channel) => channel.send("Your Frozen 🥶"))
+						.catch((err) => console.log(`${err.message} ${author}`));
 
-          return message.delete();
-        }
-      }
-    });
-  }
+					return message.delete();
+				}
+			}
+		});
+	}
 }
